Update 200514_count_PB.pl
Code was updated such that blat search is only activated if Perl cannot find a perfect of between prebarcode and prespacer sequences in sequencing reads. This reduces I/O and blast involvement, speeding up the process considerably.

It was confirmed that on two different samples, each with more than 25,000 reads, the new code and the old code produce identical results.